pyproject.toml
[project]
name = "snowflake-cs-agents"
version = "0.1.0"
description = "Intelligent Sales Assistant to get sale and conversations data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.45.1",
    "snowflake-connector-python",
    "snowflake-snowpark-python",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.12.13",
    "strands-agents[openai]>=0.3.0",
    "strands-agents-tools>=0.1.9",
    "nest-asyncio>=1.6.0",
    "pyyaml>=6.0.2",
    "orchestration-framework>=1.1.1",
]
